+++ Photo Club Hub/Extensions/PersistenceController.swift
@@ -80,7 +80,6 @@ struct PersistenceController {
             let photographer = Photographer.findCreateUpdate(
                 context: viewContext, // on main thread
                 personName: PersonName(givenName: "Jan", infixName: "D'", familyName: "Eau\(index)"),
-                organization: organization, // only needed for debug messages
                 isDeceased: memberRolesAndStatus.isDeceased(),
                 optionalFields: PhotographerOptionalFields(
                     bornDT: Date() - Double.random(in: 365*24*3600 ... 75*365*24*3600),