+++ Photo Club Hub/Documentation/ReleaseNotes.md
@@ -7,6 +7,7 @@ Code Maintenance
 * SwiftJSON imported as package (again). Was temporarily a single .swift file.
 * Entered trader status (not a trader) in Apple App Store.
 * Minor update to OrganizationType.swift (to stay in sync with Photo-Club-Hub-HTML)
+* removed Organization parameter from Photographer.createFindUpdate() as it was used for debugging only
 
 Data Maintenance
 * Added website for member WH in fgDeGender and added a few birthdays.